% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRiddles.R
\name{addRiddles}
\alias{addRiddles}
\title{addRiddles
This function allows you to add your own riddles to the existing riddles.csv file}
\usage{
addRiddles(riddle, answer, file_path = "riddles.csv")
}
\arguments{
\item{riddle}{New riddle to add to file.}

\item{answer}{New answer to add to file.}

\item{file_path}{File path for your riddles excel/csv file.}
}
\description{
addRiddles
This function allows you to add your own riddles to the existing riddles.csv file
}
\examples{
addRiddles("What has to be broken before you use it?", "egg")
}
