% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRiddles.R
\name{getRiddles}
\alias{getRiddles}
\title{playRiddles
This function allows you to add the riddles to your environment.
Riddles from https://github.com/crawsome/PyRPG_Mini.git}
\usage{
getRiddles(file_path = "data/riddles.csv")
}
\arguments{
\item{file_path}{File path for preloaded riddles excel/csv file.}
}
\description{
playRiddles
This function allows you to add the riddles to your environment.
Riddles from https://github.com/crawsome/PyRPG_Mini.git
}
\examples{
riddles <- getRiddles()
}
