% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howmanyRiddles.R
\name{howmanyRiddles}
\alias{howmanyRiddles}
\title{howmanyRiddles
This function tells you how many riddles you have.}
\usage{
howmanyRiddles(riddles)
}
\arguments{
\item{riddles}{The riddles that you get from external csv file that you add to 'riddles' object.}
}
\description{
howmanyRiddles
This function tells you how many riddles you have.
}
\examples{
howmanyRiddles(riddles)
}
