% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riddlemethis.R
\name{riddlemethis}
\alias{riddlemethis}
\title{riddlemethis
This function plays the riddle game.}
\usage{
riddlemethis(riddles)
}
\arguments{
\item{riddles}{The riddles that you get from external csv file that you add to 'riddles' object.}
}
\description{
riddlemethis
This function plays the riddle game.
}
\examples{
riddlemethis(riddles)
}
